backend/requirements.txt
# KidsChatApp Backend Dependencies
# Python >= 3.10 推薦（Python 3.14 已測試可用）

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.9.0

# Gemini SDK
google-generativeai>=0.8.0

# 選用：若需要從 .env 檔載入環境變數
# python-dotenv==1.0.1
